requirements.txt
flask
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
python-dotenv
requests
PyPDF2
pdfplumber
python-pptx
openai
groq
gunicorn
google-genai